refactor: update rust check ci
